sudoers: add audit session as timestamp_type
This implements a timestamp_type based on Linux' audit session id which
is available at /proc/<pid>/sessionid if CONFIG_AUDIT is enabled.
The concept is a middle way between "global" and "ppid". The timestamp
based on this session id is valid for all processes started under the
same audit session. This allows the timestamp being valid in all ssh
sessions when using a ControlMaster. Any subsequent ssh connections with
the same ControlMaster share the same audit session id.